requirements.txt
Django==1.9.4
djangorestframework==3.3.3
django-webpack-loader==0.2.4
wheel==0.26.0
pandas==0.19.2
django-content-editor==1.1.1
